Make Factors variables final and improve multi-threading in PrimiApp
The variable declarations inside the Factors class have been updated to 'final'. This change is necessary as these variables are not meant to be changed once initialized.
In PrimiApp, the logic has been updated to divide the number range checking for primes into smaller sections and assigning these sections to multiple threads. This is done to optimize the prime checking process, especially for big ranges of numbers. It allows the use of multi-core capabilities of modern CPUs. A Chrono is also added to track the elapsed time.