Increase range limits in PrimiApp and improve prime counting
The variable that determines the range of numbers to check for primes in PrimiApp has been adjusted from 100,000,000 to 1,000,000,000. Additionally, the ability to keep track of the total primes present and time taken to calculate them has been implemented. This not only expands the functionality of the application but allows for better benchmarking and optimization in the future by monitoring the performance of the application. More code comments have also been added for better readability.